Move editCellsMachine into subfolder
