+++ visualizer/src/Project/service/edit/editCells/index.js
@@ -1,0 +1,3 @@
+import createEditCellsMachine from './editCellsMachine';
+
+export default createEditCellsMachine;
